requirements.txt
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==3.0.0
asttokens==3.0.1
bangla-text-renderer==1.4.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.1
feedparser==6.0.12
Flask==3.1.3
flask-cors==6.0.2
Flask-WTF==1.2.2
fonttools==4.62.1
grapheme==0.6.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imgkit==1.2.3
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycairo==1.29.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.11.0
python-jose[cryptography]==3.3.0
pymongo==4.16.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
requests==2.33.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uharfbuzz==0.53.6
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.6
WTForms==3.2.1
